perf(): 修改redis限流注解最大请求数类型

+++ sakura-cache/src/main/java/com/sakura/common/cache/ratelimit/RateLimiter.java
@@ -12,17 +12,17 @@ import java.util.concurrent.TimeUnit;
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
 public @interface RateLimiter {
-    long DEFAULT_REQUEST = 10;
+    int DEFAULT_REQUEST = 10;
 
     /**
      * max 最大请求数
      */
-    @AliasFor("max") long value() default DEFAULT_REQUEST;
+    @AliasFor("max") int value() default DEFAULT_REQUEST;
 
     /**
      * max 最大请求数
      */
-    @AliasFor("value") long max() default DEFAULT_REQUEST;
+    @AliasFor("value") int max() default DEFAULT_REQUEST;
 
     /**
      * 限流key
